Refactors error logging code.
The plugin previosuly relied on `mysql_error` which was deprecated in
5.5 and completely removed in 7 in favor of the MySQL Improved
extension. Regardless, WordPress provides a method of retrieving the
last MySQL error that is interface agnostic, `$wpdb->last_error`. By
using this functionality, the plugin no longer needs to be concerned
with implementation details instead delegating that responsibility to
WordPress Core.